More map length and apple
The map has been extended, the platform is connected to the ground though. Also an apple object has been added with an animation that probably should be slowed down. It is not collectable yet.  Also the apple is the one used for pixel adventure asset set and could be changed if wanted.