Apples and Level Length
The level length is doubled. There are a few new platforms that can be changed at anyones discretion. There are also 6 apples and an apple counter that works. There is also a folder called prefabs. Those are saved objects that you can place as many times as you want without needing to do new code.