Parallax Background only moves on X-Axis and Jump Changes
In order to not make the parallax background move in the Y direction, I created a new object called "background circle". This is an invisible circle that the background is attached to instead of the player. It has a script called "followPlayerXaxis.cs" that makes it only follow the players X axis and maintain the same Y axis.

Also the Gravity is now set to 3, jumpForce set to 9.
The platform with the 2nd apple closest to spawn has been moved down one unit, so that the jump is possible with these settings.