Moving Enemy Entity
A moving object with ledge and object detection.
Both work by casting a ray vertically and horizontally to check if there is ground or an object there. If there is no ground detected in front of the entity then it turns 180 degrees the direction it was moving. If there is an object in the obstacle layer in front of the entity then it also turns 180 degrees.