Slimeball now can switch from states
So I added that the enemy can detect when the player is around him. When the player is in the right distance he will stop moving and play an animation until the player leaves. Also organized the enemy scripts (now instead of PatrolScript.cs there is a main controller called enemy.cs, patrolState.cs, PlayerDetectedState.cs and EnemyBaseState.cs) so it is easier to add more states to the enemy.