ChargeState
When the enemy with charge state detects you, after it plays the detect animation. it will then move at you at a different speed that you can set. Right now the charge speed is at 8 and the regular speed is 4. When you leave the player detection ray distance then it will reset to baseState. If you are still in the distance after a set charge time it will charge again.